feat(chart): serve device charts from the shaped tables where they cover
DeviceChartData now builds its series from device_samples and
device_state_events when both cover the requested window, and falls back
to the check-in path otherwise — no flag day, and no waiting for the
backfill to finish. A wide window becomes fixed-width rows on a clustered
index instead of tens of thousands of jsonb snapshots, each detoasted to
read four numbers out of it.

Both halves are required before the shaped path is used. Serving a window
the events do not reach would draw the numbers correctly and silently
drop the pad markers and charge runs, which is worse than being slow: a
diff on real history shows exactly that, temp and ram byte-identical
while charge collapsed to a single unknown run.

Proven by rendering the same window from both sources and comparing the
JSON. With both halves present the payloads are byte-identical, 986 bytes
each, charge runs and wlc markers included.

That diff also caught the storage type. REAL round-trips the current
firmware's readings exactly, because they originate as a Java float32 —
but a client sending a plain 31.4 came back as 31.399999618530273, which
would have put a different number on the chart than the CSV prints for
the same reading. temp_c and storage_free_gb are DOUBLE PRECISION now:
four more bytes, exact for any json number, no disagreement possible.

+++ internal/dashboard/shaped_series_test.go
@@ -10,7 +10,7 @@ import (
 
 func i16(v int16) *int16     { return &v }
 func i32(v int32) *int32     { return &v }
-func f32(v float32) *float32 { return &v }
+func f64(v float64) *float64 { return &v }
 
 // TestMergeShapedSeriesCarriesStateForward is the core of reading transitions back: a
 // state is written once and must then apply to every sample until the next event. The
@@ -21,10 +21,10 @@ func TestMergeShapedSeriesCarriesStateForward(t *testing.T) {
 	at := func(m int) time.Time { return t0.Add(time.Duration(m) * time.Minute) }
 
 	samples := []db.DeviceSample{
-		{At: at(0), BatteryPct: i16(80), TempC: f32(30.2)},
-		{At: at(5), BatteryPct: i16(79), TempC: f32(30.5)},
-		{At: at(10), BatteryPct: i16(81), TempC: f32(31.1)},
-		{At: at(15), BatteryPct: i16(83), TempC: f32(31.8)},
+		{At: at(0), BatteryPct: i16(80), TempC: f64(30.2)},
+		{At: at(5), BatteryPct: i16(79), TempC: f64(30.5)},
+		{At: at(10), BatteryPct: i16(81), TempC: f64(31.1)},
+		{At: at(15), BatteryPct: i16(83), TempC: f64(31.8)},
 	}
 	events := []db.StateAt{
 		{At: at(7), Key: "charging", Value: "true"},
